+++ src/video/sdl/sdl_window.cpp
@@ -57,18 +57,18 @@ SDLWindow::~SDLWindow()
   SDL_DestroyWindow(m_window);
 }
 
-unsigned SDLWindow::width() const
+int SDLWindow::width() const
 {
   int w;
   SDL_GetWindowSize(m_window, &w, NULL);
 
-  return static_cast<unsigned>(w);
+  return w;
 }
 
-unsigned SDLWindow::height() const
+int SDLWindow::height() const
 {
   int h;
   SDL_GetWindowSize(m_window, NULL, &h);
 
-  return static_cast<unsigned>(h);
+  return h;
 }
